Keep model-related functions together
