+++ src/CMakeLists.template.txt
@@ -41,6 +41,7 @@ set(SRC_LIST_LIB TPL_SOURCES
 ${MODEL_DIR}/wrapfunctions.cpp
 ${AMICI_DIR}/src/symbolic_functions.cpp
 ${AMICI_DIR}/src/amici_interface_cpp.cpp
+${AMICI_DIR}/src/amici_model_functions.cpp
 ${AMICI_DIR}/src/amici.cpp
 ${AMICI_DIR}/src/amici_misc.cpp
 ${AMICI_DIR}/src/udata.cpp
